+++ src/pointcloudset/io/__init__.py
@@ -7,7 +7,6 @@ from pointcloudset.io.pointcloud.csv import read_csv, write_csv
 from pointcloudset.io.pointcloud.las import read_las, write_las
 from pointcloudset.io.pointcloud.pandas import from_dataframe, to_dataframe
 from pointcloudset.io.pointcloud.pcd import read_pcd, write_pcd
-from pointcloudset.io.pointcloud.pyntcloud import from_pyntcloud, to_pyntcloud
 from pointcloudset.io.pointcloud.xyz import read_xyz, write_xyz
 
 DATASET_FROM_FILE = {
@@ -25,12 +24,10 @@ POINTCLOUD_FROM_FILE = {"CSV": read_csv, "LAS": read_las, "PCD": read_pcd, "XYZ"
 POINTCLOUD_TO_FILE = {"CSV": write_csv, "LAS": write_las, "PCD": write_pcd, "XYZ": write_xyz}
 
 POINTCLOUD_FROM_INSTANCE = {
-    "PYNTCLOUD": from_pyntcloud,
     "DATAFRAME": from_dataframe,
     "PANDAS": from_dataframe,
 }
 POINTCLOUD_TO_INSTANCE = {
-    "PYNTCLOUD": to_pyntcloud,
     "DATAFRAME": to_dataframe,
     "PANDAS": to_dataframe,
 }